requirements.txt
pillow
numpy
imageio
flake8
isort